backend/flask-backend/app/__init__.py
"""
HealthFit AI - Flask Application Factory - CLEAN ✅
"""
from flask import Flask, jsonify
from flask_sqlalchemy import SQLAlchemy
from flask_jwt_extended import JWTManager
from flask_cors import CORS

db = SQLAlchemy()
jwt = JWTManager()

def create_app(config_name: str = None) -> Flask:
    app = Flask(__name__)
    
    from app.config import Config
    app.config.from_object(Config)
    
    db.init_app(app)
    jwt.init_app(app)
    CORS(app)
    
    # JWT error handlers
    @jwt.unauthorized_loader
    def unauthorized_callback(error):
        return jsonify({"error": "Missing or invalid token"}), 401
    
    # Routes
    from app.routes.auth_routes import auth_bp
    from app.routes.health_routes import health_bp
    from app.routes.appointment_routes import appointment_bp
    from app.routes.nutrition_routes import nutrition_bp
    from app.routes.fitness_routes import fitness_bp
    from app.routes.ai_routes import ai_bp
    
    app.register_blueprint(auth_bp, url_prefix='/api/auth')
    app.register_blueprint(health_bp, url_prefix='/api/health')
    app.register_blueprint(appointment_bp, url_prefix='/api/appointments')
    app.register_blueprint(nutrition_bp, url_prefix='/api/nutrition')
    app.register_blueprint(fitness_bp, url_prefix='/api/fitness')
    app.register_blueprint(ai_bp, url_prefix='/api/ai')
    
    # Smartwatch (disabled - add file later)
    # from app.routes.smartwatch_integration import smartwatch_bp
    # app.register_blueprint(smartwatch_bp)
    
    @app.route('/api/health')
    def health_check():
        return jsonify({"status": "healthy", "ai_risk_fixed": true})
    
    with app.app_context():
        db.create_all()
    
    return app

