backend/flask-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
fitbit-api==0.3.1
requests-oauthlib==1.3.1
Werkzeug==2.3.7
